+++ src/tuple.rs
@@ -1,27 +1,27 @@
 use std::ops::{Add, Sub, Neg, Div, Mul};
 use std::cmp::PartialEq;
-use std::f32;
+use std::f64;
 use approx::{RelativeEq, AbsDiffEq};
 
 
 #[derive(Debug, PartialEq, PartialOrd, Clone, Copy)]
 pub struct Tuple {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
-    pub w: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
+    pub w: f64,
 }
 
 impl Tuple {
-    pub fn new(x: f32, y: f32, z: f32, w: f32) -> Self {
+    pub fn new(x: f64, y: f64, z: f64, w: f64) -> Self {
         Tuple{ x, y, z, w }
     }
 
-    pub fn from_array(arr: [f32; 4]) -> Self {
+    pub fn from_array(arr: [f64; 4]) -> Self {
         Tuple{ x: arr[0], y: arr[1], z: arr[2], w: arr[3] }
     }
 
-    pub fn dot(&self, rhs: Tuple) -> f32 {
+    pub fn dot(&self, rhs: Tuple) -> f64 {
         self.x * rhs.x + self.y * rhs.y + self.z * rhs.z + self.w * rhs.w
     }
 
@@ -34,7 +34,7 @@ impl Tuple {
         }
     }
 
-    pub fn abs(&self) -> f32 {
+    pub fn abs(&self) -> f64 {
         (self.x.powi(2) + self.y.powi(2) + self.z.powi(2) + self.w.powi(2)).sqrt()
     }
 }
@@ -78,10 +78,10 @@ impl Neg for Tuple {
     }
 }
 
-impl Mul<f32> for Tuple {
+impl Mul<f64> for Tuple {
     type Output = Self;
 
-    fn mul(self, rhs: f32) -> Self::Output {
+    fn mul(self, rhs: f64) -> Self::Output {
         Self {
             x: self.x * rhs,
             y: self.y * rhs,
@@ -91,10 +91,10 @@ impl Mul<f32> for Tuple {
     }
 }
 
-impl Div<f32> for Tuple {
+impl Div<f64> for Tuple {
     type Output = Self;
 
-    fn div(self, rhs: f32) -> Self::Output {
+    fn div(self, rhs: f64) -> Self::Output {
         Self {
             x: self.x / rhs,
             y: self.y / rhs,
@@ -105,31 +105,31 @@ impl Div<f32> for Tuple {
 }
 
 impl AbsDiffEq for Tuple {
-    type Epsilon = f32;
+    type Epsilon = f64;
 
     fn default_epsilon() -> Self::Epsilon {
-        f32::default_epsilon()
+        f64::default_epsilon()
     }
 
-    fn abs_diff_eq(&self, other: &Self, epsilon: f32) -> bool {
-        f32::abs_diff_eq(&self.x, &other.x, epsilon) &&
-        f32::abs_diff_eq(&self.y, &other.z, epsilon) &&
-        f32::abs_diff_eq(&self.z, &other.z, epsilon) &&
-        f32::abs_diff_eq(&self.w, &other.w, epsilon)
+    fn abs_diff_eq(&self, other: &Self, epsilon: f64) -> bool {
+        f64::abs_diff_eq(&self.x, &other.x, epsilon) &&
+        f64::abs_diff_eq(&self.y, &other.z, epsilon) &&
+        f64::abs_diff_eq(&self.z, &other.z, epsilon) &&
+        f64::abs_diff_eq(&self.w, &other.w, epsilon)
     }
 }
 
 impl RelativeEq for Tuple {
 
-    fn default_max_relative() -> f32 {
-        f32::default_max_relative()
+    fn default_max_relative() -> f64 {
+        f64::default_max_relative()
     }
 
-    fn relative_eq(&self, other: &Self, epsilon: f32, max_relative: f32) -> bool {
-        f32::relative_eq(&self.x, &other.x, epsilon, max_relative) &&
-        f32::relative_eq(&self.y, &other.y, epsilon, max_relative) &&
-        f32::relative_eq(&self.z, &other.z, epsilon, max_relative) &&
-        f32::relative_eq(&self.w, &other.w, epsilon, max_relative)
+    fn relative_eq(&self, other: &Self, epsilon: f64, max_relative: f64) -> bool {
+        f64::relative_eq(&self.x, &other.x, epsilon, max_relative) &&
+        f64::relative_eq(&self.y, &other.y, epsilon, max_relative) &&
+        f64::relative_eq(&self.z, &other.z, epsilon, max_relative) &&
+        f64::relative_eq(&self.w, &other.w, epsilon, max_relative)
     }
 }
 
@@ -139,7 +139,7 @@ pub struct TupleIntoIterator {
 }
 
 impl IntoIterator for Tuple {
-    type Item = f32;
+    type Item = f64;
     type IntoIter = TupleIntoIterator;
 
     fn into_iter(self) -> Self::IntoIter {
@@ -151,8 +151,8 @@ impl IntoIterator for Tuple {
 }
 
 impl Iterator for TupleIntoIterator {
-    type Item = f32;
-    fn next(&mut self) -> Option<f32> {
+    type Item = f64;
+    fn next(&mut self) -> Option<f64> {
         let result = match self.index {
             0 => self.tuple.x,
             1 => self.tuple.y,
